Added and fixed some static function declarations in xreffile.c Fixed a cast in translat.c

+++ src/gedlib/translat.c
@@ -164,7 +164,7 @@ translate_string (TRANTABLE tt, /* tran table */
 	n = strlen(in);
 	l = 0;
 	while (n > 0) {
-		node = tt->start[*p];
+		node = tt->start[(int)*p];
 		if (!node) {	/* this char starts no patterns */
 			add_char(out, &l, max, *p++);
 			--n;
